tweet_collector/requirements.txt
tweepy
pandas
sqlalchemy
psycopg2-binary
python-dotenv
vaderSentiment
